+++ cave/utils/exceptions.py
@@ -1,0 +1,8 @@
+class Deactivated(Exception):
+    pass
+
+class NotApplicable(Exception):
+    pass
+
+class NotUniqueError(Exception):
+    pass
