+++ cave/utils/options/default_bohb_analysis_options.ini
@@ -1,0 +1,75 @@
+[DEFAULT]
+run = True
+
+[Meta Data]
+
+[Ablation]
+
+[Algorithm Footprint]
+run=False
+
+[Best Configuration]
+
+[Budget Correlation]
+
+[BOHB Learning Curves]
+
+[Configurator Footprint]
+# whether or not to have a time_slider-widget on cfp-plot
+# INCREASES FILE-SIZE DRAMATICALLY
+time_slider = True
+# maximum number of data-points to plot
+max_configurations_to_plot = -1
+# if use_timeslider is not off, defines the number of quantiles for the
+# slider/ number of static pictures
+number_quantiles = 10
+# whether to use a logarithmic scale for the timeslider/quantiles
+timeslider_log = True
+
+[Cost Over Time]
+# from ['racing', 'minimum', 'prefer_higher_budget'], defines incumbent trajectory from hpbandster result
+incumbent_trajectory = racing
+# if True, average over plots. if False, all runs are treated individually with checkboxes
+average_over_runs = True
+
+[empirical Cumulative Distribution Function (eCDF)]
+run=False
+
+[fANOVA]
+fanova_pairwise = True
+pimp_max_samples = -1
+
+[Feature Clustering]
+run=False
+
+[Feature Correlation]
+run=False
+
+[Feature Importance]
+run=False
+
+[Forward Selection]
+run=False
+
+[Importance Table]
+sort_table_by = average
+
+[Incumbents Over Budgets]
+
+[Local Parameter Importance (LPI)]
+
+[Parallel Coordinates]
+pc_sort_by = none
+params = 5
+n_configs = 100
+max_runs_epm = 300000
+
+[Performance Table]
+run=False
+
+[Scatterplot]
+run=False
+
+[Violin and Box Plots]
+run=False
+
