+++ setup.py
@@ -31,7 +31,8 @@ setuptools.setup(
             'cave/plot/mpl_style'
         ],
         'cave/utils' : [
-            'cave/utils/default_analysis_options.ini'
+            'cave/utils/options/default_analysis_options.ini'
+            'cave/utils/options/default_bohb_analysis_options.ini'
         ],
         'cave/html': [
             'cave/html/web_files/css/accordion.css',
